Rework GCF entrypoint
It now uses Google's officially recommended way to hook the function via
`@google-cloud/functions-framework` layer.

This should fix silent hangs we experienced since upgrading to Probot v13
