--- function.js
@@ -1,4 +1,0 @@
-const {createNodeMiddleware, createProbot} = require('probot')
-const app = require('./lib/index.js')
-
-exports.probotApp = createNodeMiddleware(app, { probot: createProbot() })
